feature: AStarController
Trouxe os arquivos da main para uma classe na pasta controller. A main será adicionada posteriormente como gerenciador de rotas
