--- src/FlightWise/src/main/java/com/example/mdbspringboot/MdbSpringBootApplication.java
@@ -1,22 +1,0 @@
-package com.example.mdbspringboot;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@SpringBootApplication
-@EnableMongoRepositories
-public class MdbSpringBootApplication implements CommandLineRunner {
-    @Autowired
-    RouteRepository routeRepository;
-    public static void main(String[] args) {
-        SpringApplication.run(MdbSpringBootApplication.class, args);
-    }
-
-    @Override
-    public void run(String... args) throws Exception {
-
-    }
-}
